@@ -1,7 +1,7 @@
 name: seatsio
 description: Seatsio SDK for Flutter.
 repository: https://github.com/SongJiaqiang/seatsio-flutter
-version: 0.3.4
+version: 0.4.0
 
 environment:
   sdk: '>=3.0.6 <4.0.0'
@@ -11,23 +11,14 @@ dependencies:
   flutter:
     sdk: flutter
 
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   built_collection: ^5.1.1
-  built_value: ^8.7.0
-  webview_flutter: ^4.4.2
+  built_value: ^8.9.3
+  webview_flutter: ^4.10.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^2.4.6
-  built_value_generator: ^8.7.0
-
-# The following section is specific to Flutter.
-flutter:
-
-  # To add assets to your package, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  #
+  build_runner: ^2.4.15
+  built_value_generator: ^8.9.3
