@@ -7,9 +7,10 @@ export 'src/models/seating_chart_config.dart';
 export 'src/models/seating_chart.dart';
 export 'src/models/seating_config_change.dart';
 export 'src/models/seatsio_category.dart';
-export 'src/models/seatsio_object.dart';
+export 'src/models/seatsio_config_colors.dart';
 export 'src/models/seatsio_label.dart';
 export 'src/models/pricing_for_category.dart';
+export 'src/models/seatsio_object.dart';
 export 'src/models/seatsio_point.dart';
 export 'src/models/seatsio_ticket_type.dart';
 export 'src/models/hold_token.dart';
