@@ -4,7 +4,7 @@ final String seatsioHTML = """
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <script src="https://cdn-%region%.seatsio.net/chart.js" type="text/javascript"></script>
   </head>
-    <body style="margin: 0; padding: 0">
+    <body style="margin: 0; padding: 0;">
       <div id="chart" style="width: 100%; height: 100%;"></div>
       <script>
         let chart = new seatsio.SeatingChart(%configAsJs%).render()
@@ -18,6 +18,10 @@ final String seatsioHTML = """
         function changeConfig(data, callback) {
             chart.changeConfig(JSON.parse(data))
             callback(data)
+        }
+
+        function deselectObjects(data) {
+            chart.deselectObjects(JSON.parse(data))
         }
 
         // [end]
