@@ -41,7 +41,8 @@ class _$PricingForCategorySerializer
     if (value != null) {
       result
         ..add('price')
-        ..add(serializers.serialize(value, specifiedType: const FullType(num)));
+        ..add(serializers.serialize(value,
+            specifiedType: const FullType(String)));
     }
     value = object.ticketTypes;
     if (value != null) {
@@ -76,7 +77,7 @@ class _$PricingForCategorySerializer
           break;
         case 'price':
           result.price = serializers.deserialize(value,
-              specifiedType: const FullType(num)) as num?;
+              specifiedType: const FullType(String)) as String?;
           break;
         case 'ticketTypes':
           result.ticketTypes.replace(serializers.deserialize(value,
@@ -163,7 +164,7 @@ class _$PricingForCategory extends PricingForCategory {
   @override
   final String? category;
   @override
-  final num? price;
+  final String? price;
   @override
   final BuiltList<TicketTypePricing>? ticketTypes;
 
@@ -228,9 +229,9 @@ class PricingForCategoryBuilder
   String? get category => _$this._category;
   set category(String? category) => _$this._category = category;
 
-  num? _price;
-  num? get price => _$this._price;
-  set price(num? price) => _$this._price = price;
+  String? _price;
+  String? get price => _$this._price;
+  set price(String? price) => _$this._price = price;
 
   ListBuilder<TicketTypePricing>? _ticketTypes;
   ListBuilder<TicketTypePricing> get ticketTypes =>
