@@ -1,5 +1,6 @@
 import 'dart:convert';
-import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
+import 'package:seatsio/src/util/constants.dart';
 import '../util/seatsio_web_view_controller.dart';
 import 'seating_config_change.dart';
 import 'seatsio_category.dart';
@@ -19,8 +20,22 @@ class SeatingChart {
     // inject javascript to SeatsioWebView
   }
 
-  void deselectObject(List<String> objects) {
-    // inject javascript to SeatsioWebView
+  Future<void> deselectObject(List<String> objects) async {
+    if (objects.isEmpty) {
+      kDebugPrint("[Seatsio]-> deselectObject: No objects provided.");
+      return;
+    }
+
+    final objectsJson = jsonEncode(objects);
+    final jsString = "deselectObjects('$objectsJson');";
+
+    try {
+      seatsioController.evaluateJavascript(jsString);
+    } on PlatformException catch (error) {
+      kDebugPrint("[Seatsio]-> deselectObject PlatformException: $error");
+    } catch (error) {
+      kDebugPrint("[Seatsio]-> Unknown error in deselectObject: $error");
+    }
   }
 
   void changeConfig(SeatingConfigChange configChange) {
@@ -29,8 +44,13 @@ class SeatingChart {
     configJson = configJson.replaceAll("'", "\\\'");
 
     final jsString = "changeConfig('$configJson', postMessageToFlutter)";
-    debugPrint("[Seatsio]-> changeConfig jsString: $jsString");
-    seatsioController.evaluateJavascript(jsString);
+    kDebugPrint("[Seatsio]-> changeConfig jsString: $jsString");
+
+    try {
+      seatsioController.evaluateJavascript(jsString);
+    } catch (error) {
+      kDebugPrint("[Seatsio]-> changeConfig error: $error");
+    }
   }
 
   void listCategories(CategoryListCallback callback) {
@@ -38,7 +58,12 @@ class SeatingChart {
     // get category from bridge callback
 
     final jsString = "listCategories(null, postMessageToFlutter);";
-    seatsioController.evaluateJavascript(jsString);
+
+    try {
+      seatsioController.evaluateJavascript(jsString);
+    } catch (error) {
+      kDebugPrint("[Seatsio]-> listCategories error: $error");
+    }
 
     callback([]);
   }
@@ -48,7 +73,12 @@ class SeatingChart {
     // get category from bridge callback
 
     final jsString = "listCategories(null, postMessageToFlutter);";
-    seatsioController.evaluateJavascript(jsString);
+
+    try {
+      seatsioController.evaluateJavascript(jsString);
+    } catch (error) {
+      kDebugPrint("[Seatsio]-> listCategories error: $error");
+    }
   }
 
   @override
