@@ -4,6 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:seatsio/src/models/hold_token.dart';
+import 'package:seatsio/src/util/constants.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
 import '../models/seating_chart.dart';
@@ -21,6 +22,7 @@ class SeatsioWebView extends StatefulWidget {
     SeatsioCategoryListCallback? onCategoryListCallback,
     SeatingChartCallback? onChartRendered,
     VoidCallback? onChartRenderingFailed,
+    VoidCallback? onChartRerenderingStarted,
     SeatsioObjectCallback? onObjectClicked,
     SeatsioObjectTicketTypeCallback? onObjectSelected,
     SeatsioObjectTicketTypeCallback? onObjectDeselected,
@@ -42,6 +44,7 @@ class SeatsioWebView extends StatefulWidget {
         this._onCategoryListCallback = onCategoryListCallback,
         this._onChartRendered = onChartRendered,
         this._onChartRenderingFailed = onChartRenderingFailed,
+        this._onChartRerenderingStarted = onChartRerenderingStarted,
         this._onObjectClicked = onObjectClicked,
         this._onObjectSelected = onObjectSelected,
         this._onObjectDeselected = onObjectDeselected,
@@ -77,6 +80,8 @@ class SeatsioWebView extends StatefulWidget {
   final SeatingChartCallback? _onChartRendered;
 
   final VoidCallback? _onChartRenderingFailed;
+
+  final VoidCallback? _onChartRerenderingStarted;
 
   /// Fired when best available objects are successfully selected. This callback receives two parameters:
   /// [array_of_objects]: the best available objects
@@ -119,42 +124,47 @@ class _SeatsioWebViewState extends State<SeatsioWebView> {
   void initState() {
     super.initState();
 
-    _webViewController = WebViewController()
-      ..setJavaScriptMode(JavaScriptMode.unrestricted)
-      ..addJavaScriptChannel('FlutterJsBridge', onMessageReceived: flutterJsBridge)
-      ..addJavaScriptChannel('onObjectClicked', onMessageReceived: onObjectClicked)
-      ..addJavaScriptChannel('onObjectSelected', onMessageReceived: onObjectSelected)
-      ..addJavaScriptChannel('onObjectDeselected', onMessageReceived: onObjectDeselected)
-      ..addJavaScriptChannel('onChartRendered', onMessageReceived: onChartRendered)
-      ..addJavaScriptChannel('onChartRenderingFailed', onMessageReceived: onChartRenderingFailed)
-      ..addJavaScriptChannel('onSelectionValid', onMessageReceived: onSelectionValid)
-      ..addJavaScriptChannel('onSelectionInvalid', onMessageReceived: onSelectionInvalid)
-      ..addJavaScriptChannel('onBestAvailableSelectionFailed', onMessageReceived: onBestAvailableSelectionFailed)
-      ..addJavaScriptChannel('onHoldSucceeded', onMessageReceived: onHoldSucceeded)
-      ..addJavaScriptChannel('onHoldFailed', onMessageReceived: onHoldFailed)
-      ..addJavaScriptChannel('onHoldTokenExpired', onMessageReceived: onHoldTokenExpired)
-      ..addJavaScriptChannel('onSessionInitialized', onMessageReceived: onSessionInitialized)
-      ..addJavaScriptChannel('onReleaseHoldSucceeded', onMessageReceived: onReleaseHoldSucceeded)
-      ..addJavaScriptChannel('onReleaseHoldFailed', onMessageReceived: onReleaseHoldFailed)
-      ..setNavigationDelegate(
-        NavigationDelegate(
-          onPageFinished: (url) {
-            if (widget._enableDebug) debugPrint("[Seatsio]-> onPageFinished: $url");
-          },
-          onWebResourceError: (error) {
-            if (widget._enableDebug) debugPrint("[Seatsio]-> onWebResourceError: ${error.description}");
-          },
-          onNavigationRequest: (request) {
-            if (widget._enableDebug) debugPrint("[Seatsio]-> onNavigationRequest: ${request.url}");
-            return NavigationDecision.navigate;
-          },
-        ),
-      );
-    if (widget._initialUrl != null) {
-      _webViewController.loadRequest(Uri.parse(widget._initialUrl!));
+    try {
+      _webViewController = WebViewController()
+        ..setJavaScriptMode(JavaScriptMode.unrestricted)
+        ..addJavaScriptChannel('FlutterJsBridge', onMessageReceived: flutterJsBridge)
+        ..addJavaScriptChannel('onObjectClicked', onMessageReceived: onObjectClicked)
+        ..addJavaScriptChannel('onObjectSelected', onMessageReceived: onObjectSelected)
+        ..addJavaScriptChannel('onObjectDeselected', onMessageReceived: onObjectDeselected)
+        ..addJavaScriptChannel('onChartRendered', onMessageReceived: onChartRendered)
+        ..addJavaScriptChannel('onChartRenderingFailed', onMessageReceived: onChartRenderingFailed)
+        ..addJavaScriptChannel('onChartRerenderingStarted', onMessageReceived: onChartRerenderingStarted)
+        ..addJavaScriptChannel('onSelectionValid', onMessageReceived: onSelectionValid)
+        ..addJavaScriptChannel('onSelectionInvalid', onMessageReceived: onSelectionInvalid)
+        ..addJavaScriptChannel('onBestAvailableSelectionFailed', onMessageReceived: onBestAvailableSelectionFailed)
+        ..addJavaScriptChannel('onHoldSucceeded', onMessageReceived: onHoldSucceeded)
+        ..addJavaScriptChannel('onHoldFailed', onMessageReceived: onHoldFailed)
+        ..addJavaScriptChannel('onHoldTokenExpired', onMessageReceived: onHoldTokenExpired)
+        ..addJavaScriptChannel('onSessionInitialized', onMessageReceived: onSessionInitialized)
+        ..addJavaScriptChannel('onReleaseHoldSucceeded', onMessageReceived: onReleaseHoldSucceeded)
+        ..addJavaScriptChannel('onReleaseHoldFailed', onMessageReceived: onReleaseHoldFailed)
+        ..setNavigationDelegate(
+          NavigationDelegate(
+            onPageFinished: (url) {
+              if (widget._enableDebug) kDebugPrint("[Seatsio]-> onPageFinished: $url");
+            },
+            onWebResourceError: (error) {
+              if (widget._enableDebug) kDebugPrint("[Seatsio]-> onWebResourceError: ${error.description}");
+            },
+            onNavigationRequest: (request) {
+              if (widget._enableDebug) kDebugPrint("[Seatsio]-> onNavigationRequest: ${request.url}");
+              return NavigationDecision.navigate;
+            },
+          ),
+        );
+      if (widget._initialUrl != null) {
+        _webViewController.loadRequest(Uri.parse(widget._initialUrl!));
+      }
+      _seatsioController = SeatsioWebViewController(webViewController: _webViewController);
+      widget._onWebViewCreated?.call(_seatsioController);
+    } catch (error) {
+      kDebugPrint("[Seatsio]-> Error on init webview: $error");
     }
-    _seatsioController = SeatsioWebViewController(webViewController: _webViewController);
-    widget._onWebViewCreated?.call(_seatsioController);
   }
 
   @override
@@ -167,7 +177,7 @@ class _SeatsioWebViewState extends State<SeatsioWebView> {
 
   // Handle default callback from javascript
   void flutterJsBridge(JavaScriptMessage message) {
-    if (widget._enableDebug) debugPrint("[Seatsio]-> _onFlutterJsBridgeChannel callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> _onFlutterJsBridgeChannel callback message: ${message.message}");
     // Get categories of chart if call chart.requestCategories() on the onChartRendered callback.
     final categories = SeatsioCategory.arrayFromJson(message.message);
     if (categories != null && categories.isNotEmpty) {
@@ -177,55 +187,61 @@ class _SeatsioWebViewState extends State<SeatsioWebView> {
 
   void onObjectClicked(JavaScriptMessage message) {
     if (widget._onObjectClicked == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onObjectClicked callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onObjectClicked callback message: ${message.message}");
     final object = SeatsioObject.fromJson(message.message);
     if (object != null) {
       widget._onObjectClicked?.call(object);
     }
   }
 
-  void onObjectSelected(JavaScriptMessage message) {
+  Future<void> onObjectSelected(JavaScriptMessage message) async {
     if (widget._onObjectSelected == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onObjectClicked callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onObjectSelected callback message: ${message.message}");
     final object = SeatsioObject.fromJson(message.message);
     if (object != null) {
-      widget._onObjectSelected?.call(object, null);
+      await widget._onObjectSelected?.call(object, null);
     }
   }
 
-  void onObjectDeselected(JavaScriptMessage message) {
+  Future<void> onObjectDeselected(JavaScriptMessage message) async {
     if (widget._onObjectDeselected == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onObjectDeselected callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onObjectDeselected callback message: ${message.message}");
     final object = SeatsioObject.fromJson(message.message);
     if (object != null) {
       // todo: Miss ticketType parameter
       // Does not found ticketType from javascript message like iOS SDK.
-      widget._onObjectDeselected?.call(object, null);
+      await widget._onObjectDeselected?.call(object, null);
     }
   }
 
   void onChartRendered(JavaScriptMessage message) {
     if (widget._onChartRendered == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onChartRendered callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onChartRendered callback message: ${message.message}");
     final seatingChart = SeatingChart(_seatsioController);
     widget._onChartRendered?.call(seatingChart);
   }
 
   void onChartRenderingFailed(JavaScriptMessage message) {
     if (widget._onChartRenderingFailed == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onChartRenderingFailed callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onChartRenderingFailed callback message: ${message.message}");
     widget._onChartRenderingFailed?.call();
+  }
+
+  void onChartRerenderingStarted(JavaScriptMessage message) {
+    if (widget._onChartRerenderingStarted == null) return;
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onChartRerenderingStarted callback message: ${message.message}");
+    widget._onChartRerenderingStarted?.call();
   }
 
   void onSelectionValid(JavaScriptMessage message) {
     if (widget._onSelectionValid == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onSelectionValid callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onSelectionValid callback message: ${message.message}");
     widget._onSelectionValid?.call();
   }
 
   void onSelectionInvalid(JavaScriptMessage message) {
     if (widget._onSelectionInvalid == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onSelectionInvalid callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onSelectionInvalid callback message: ${message.message}");
     // todo: Get valid types from message
     // I don't known how to decode the SelectionValidatorType from the javascript message,
     // So I always return an empty list.
@@ -234,7 +250,7 @@ class _SeatsioWebViewState extends State<SeatsioWebView> {
 
   void onBestAvailableSelected(JavaScriptMessage message) {
     if (widget._onBestAvailableSelected == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onBestAvailableSelected callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onBestAvailableSelected callback message: ${message.message}");
     // todo: Get array_of_objects and nextToEachOther from message
     final objects = SeatsioObject.arrayFromJson(message.message);
     if (objects != null) {
@@ -247,13 +263,13 @@ class _SeatsioWebViewState extends State<SeatsioWebView> {
   void onBestAvailableSelectionFailed(JavaScriptMessage message) {
     if (widget._onBestAvailableSelectionFailed == null) return;
     if (widget._enableDebug)
-      debugPrint("[Seatsio]-> onBestAvailableSelectionFailed callback message: ${message.message}");
+      kDebugPrint("[Seatsio]-> onBestAvailableSelectionFailed callback message: ${message.message}");
     widget._onBestAvailableSelectionFailed?.call();
   }
 
   void onHoldSucceeded(JavaScriptMessage message) {
     if (widget._onHoldSucceeded == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onHoldSucceeded callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onHoldSucceeded callback message: ${message.message}");
     // todo: what about ticket types?
     final objects = SeatsioObject.arrayFromJson(message.message);
     if (objects != null) {
@@ -265,7 +281,7 @@ class _SeatsioWebViewState extends State<SeatsioWebView> {
 
   void onHoldFailed(JavaScriptMessage message) {
     if (widget._onHoldFailed == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onHoldFailed callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onHoldFailed callback message: ${message.message}");
     // todo: what about ticket types?
     final objects = SeatsioObject.arrayFromJson(message.message);
     if (objects != null) {
@@ -277,34 +293,34 @@ class _SeatsioWebViewState extends State<SeatsioWebView> {
 
   void onHoldTokenExpired(JavaScriptMessage message) {
     if (widget._onHoldTokenExpired == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onHoldTokenExpired callback was called");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onHoldTokenExpired callback was called");
     widget._onHoldTokenExpired?.call();
   }
 
   void onSessionInitialized(JavaScriptMessage message) {
     if (widget._onSessionInitialized == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onSessionInitialized callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onSessionInitialized callback message: ${message.message}");
     final holdToken = HoldToken.fromJson(jsonDecode(message.message));
     widget._onSessionInitialized?.call(holdToken);
   }
 
   void onReleaseHoldSucceeded(JavaScriptMessage message) {
     if (widget._onReleaseHoldSucceeded == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onReleaseHoldSucceeded callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onReleaseHoldSucceeded callback message: ${message.message}");
     // todo: get objects and types from message
     widget._onReleaseHoldSucceeded?.call([], null);
   }
 
   void onReleaseHoldFailed(JavaScriptMessage message) {
     if (widget._onReleaseHoldFailed == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onReleaseHoldFailed callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onReleaseHoldFailed callback message: ${message.message}");
     // todo: get objects and types from message
     widget._onReleaseHoldFailed?.call([], null);
   }
 
   void onSelectedObjectBooked(JavaScriptMessage message) {
     if (widget._onSelectedObjectBooked == null) return;
-    if (widget._enableDebug) debugPrint("[Seatsio]-> onSelectedObjectBooked callback message: ${message.message}");
+    if (widget._enableDebug) kDebugPrint("[Seatsio]-> onSelectedObjectBooked callback message: ${message.message}");
     final object = SeatsioObject.fromJson(message.message);
     if (object != null) {
       widget._onSelectedObjectBooked?.call(object);
