@@ -2,8 +2,7 @@ import '../models/seating_chart_config.dart';
 
 class SeatsioJsBridge {
   /// 将callback转成字符串
-  static List<String> buildCallbacksConfiguration(
-      SeatingChartConfig chartConfig) {
+  static List<String> buildCallbacksConfiguration(SeatingChartConfig chartConfig) {
     final List<String> callbacks = [];
 
     if (chartConfig.enableChartRenderedCallback) {
@@ -12,6 +11,10 @@ class SeatsioJsBridge {
 
     if (chartConfig.enableChartRenderingFailedCallback) {
       callbacks.add(buildCallbackConfigAsJS("onChartRenderingFailed"));
+    }
+
+    if (chartConfig.enableChartRerenderingStartedCallback) {
+      callbacks.add(buildCallbackConfigAsJS("onChartRerenderingStarted"));
     }
 
     if (chartConfig.enableObjectClickedCallback) {
@@ -78,6 +81,12 @@ class SeatsioJsBridge {
   }
 
   static String seatsioInjectString(String selectedFeature) {
-    return "(object, dfValue, extraConfig) => {if(extraConfig[object.label] == 'true') {return '$selectedFeature'} else {return dfValue}}";
+    return "(object, dfValue, extraConfig) => {"
+        "if(extraConfig[object.label] == 'true') {"
+        "return '$selectedFeature'"
+        "} else {"
+        "return dfValue"
+        "}"
+        "}";
   }
 }
