@@ -1,10 +1,11 @@
 import 'dart:convert';
+import 'dart:developer';
 
-import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:seatsio/src/assets/seatsio_html.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 import '../models/seating_chart_config.dart';
 import '../util/seatsio_js_bridge.dart';
-import '../assets/seatsio_html.dart';
 
 typedef void SeatsioWebViewCreatedCallback(SeatsioWebViewController controller);
 
@@ -13,32 +14,42 @@ class SeatsioWebViewController {
     required WebViewController webViewController,
   }) : this._webViewController = webViewController;
 
-  WebViewController _webViewController;
+  final WebViewController _webViewController;
 
   SeatingChartConfig? _chartConfig;
 
   void reload(SeatingChartConfig? newChartConfig) {
     if (newChartConfig != null) {
       _chartConfig = newChartConfig;
+    } else {
+      return;
     }
 
     if (_chartConfig != null) {
       final url = _generateHtmlContent(_chartConfig!);
-      _webViewController.loadRequest(Uri.parse(url));
+      _webViewController
+        ..setBackgroundColor(Colors.transparent)
+        ..loadRequest(Uri.parse(url));
     } else {
-      debugPrint("[Seatsio]-> Not found seatsio chart config info.");
+      log("[Seatsio]-> Not found seatsio chart config info.");
     }
   }
 
-  void reloadChart(String token, String session) {
+  void reloadChart({required String token, required String session}) {
     if (_chartConfig != null) {
-      final newChartInfo = _chartConfig!.rebuild((b) => b
-        ..holdToken = token
-        ..session = session);
+      final newChartInfo = _chartConfig!.rebuild(
+        (b) => b
+          ..holdToken = token
+          ..session = session,
+      );
       final url = _generateHtmlContent(newChartInfo);
-      _webViewController.loadRequest(Uri.parse(url));
+      try {
+        _webViewController.loadRequest(Uri.parse(url));
+      } catch (e) {
+        log("[Seatsio]-> Error while reloading chart: $e");
+      }
     } else {
-      debugPrint("[Seatsio]-> Not found seatsio chart config info.");
+      log("[Seatsio]-> Not found seatsio chart config info.");
     }
   }
 
@@ -48,22 +59,19 @@ class SeatsioWebViewController {
     final chartConfigMap = chartConfig.toMap();
 
     // Convert map to json string
-    String chartConfigJson = jsonEncode(chartConfigMap);
-    chartConfigJson = '$chartConfigJson';
+    var chartConfigJson = jsonEncode(chartConfigMap);
 
     // Append callback string to json string.
     final callbacks = SeatsioJsBridge.buildCallbacksConfiguration(chartConfig);
-    chartConfigJson = chartConfigJson.substring(0, chartConfigJson.length - 1);
-    callbacks.forEach((e) {
-      chartConfigJson = "$chartConfigJson, $e";
-    });
-    chartConfigJson = "$chartConfigJson}";
+    final sb = StringBuffer(chartConfigJson.substring(0, chartConfigJson.length - 1));
+    callbacks.forEach((e) => sb.write(', $e'));
+    sb.write('}');
+    chartConfigJson = sb.toString();
 
     // Insert json string of chart config to the seatsio HTML template.
-    final htmlString =
-        seatsioHTML.replaceFirst("%region%", chartConfig.region ?? "eu").replaceFirst("%configAsJs%", chartConfigJson);
-
-    debugPrint("[Seatsio]-> _generateHtmlContent: $htmlString");
+    final htmlString = seatsioHTML //
+        .replaceFirst("%region%", chartConfig.region ?? "eu")
+        .replaceFirst("%configAsJs%", chartConfigJson);
 
     // Encode HTML string with utf8
     final url = Uri.dataFromString(
@@ -75,5 +83,7 @@ class SeatsioWebViewController {
     return url.toString();
   }
 
-  Future<void> evaluateJavascript(String javascriptString) => _webViewController.runJavaScript(javascriptString);
+  Future<void> evaluateJavascript(String javascriptString) {
+    return _webViewController.runJavaScript(javascriptString);
+  }
 }
